refactor L2Source interface
